+++ sdk/resourcemanager/dashboard/armdashboard/constants.go
@@ -10,7 +10,7 @@ package armdashboard
 
 const (
 	moduleName    = "github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/dashboard/armdashboard"
-	moduleVersion = "v1.2.0"
+	moduleVersion = "v2.0.0"
 )
 
 type APIKey string
@@ -25,20 +25,6 @@ func PossibleAPIKeyValues() []APIKey {
 	return []APIKey{
 		APIKeyDisabled,
 		APIKeyEnabled,
-	}
-}
-
-// ActionType - Enum. Indicates the action type. "Internal" refers to actions that are for internal only APIs.
-type ActionType string
-
-const (
-	ActionTypeInternal ActionType = "Internal"
-)
-
-// PossibleActionTypeValues returns the possible values for the ActionType const type.
-func PossibleActionTypeValues() []ActionType {
-	return []ActionType{
-		ActionTypeInternal,
 	}
 }
 
@@ -159,25 +145,6 @@ func PossibleMarketplaceAutoRenewValues() []MarketplaceAutoRenew {
 	return []MarketplaceAutoRenew{
 		MarketplaceAutoRenewDisabled,
 		MarketplaceAutoRenewEnabled,
-	}
-}
-
-// Origin - The intended executor of the operation; as in Resource Based Access Control (RBAC) and audit logs UX. Default
-// value is "user,system"
-type Origin string
-
-const (
-	OriginSystem     Origin = "system"
-	OriginUser       Origin = "user"
-	OriginUserSystem Origin = "user,system"
-)
-
-// PossibleOriginValues returns the possible values for the Origin const type.
-func PossibleOriginValues() []Origin {
-	return []Origin{
-		OriginSystem,
-		OriginUser,
-		OriginUserSystem,
 	}
 }