+++ sdk/resourcemanager/dashboard/armdashboard/responses.go
@@ -89,12 +89,6 @@ type ManagedPrivateEndpointsClientUpdateResponse struct {
 	ManagedPrivateEndpointModel
 }
 
-// OperationsClientListResponse contains the response from method OperationsClient.NewListPager.
-type OperationsClientListResponse struct {
-	// A list of REST API operations supported by an Azure Resource Provider. It contains an URL link to get the next set of results.
-	OperationListResult
-}
-
 // PrivateEndpointConnectionsClientApproveResponse contains the response from method PrivateEndpointConnectionsClient.BeginApprove.
 type PrivateEndpointConnectionsClientApproveResponse struct {
 	// The Private Endpoint Connection resource.