+++ sdk/resourcemanager/dashboard/armdashboard/CHANGELOG.md
@@ -1,5 +1,18 @@
 # Release History
 
+## 2.0.0 (2024-06-13)
+### Breaking Changes
+
+- Enum `ActionType` has been removed
+- Enum `Origin` has been removed
+- Function `*ClientFactory.NewOperationsClient` has been removed
+- Function `NewOperationsClient` has been removed
+- Function `*OperationsClient.NewListPager` has been removed
+- Struct `Operation` has been removed
+- Struct `OperationDisplay` has been removed
+- Struct `OperationListResult` has been removed
+
+
 ## 1.2.0 (2023-11-24)
 ### Features Added